VarBinVector
